refactor(frontend): convert SecretsApiClient .js+.d.ts to single typed .ts (#12363)
